chore(temp file names): add temp names for the files
