+++ apps/landing/src/components/explanation.tmp.tsx
