+++ apps/landing/src/components/hero.tmp.tsx
