+++ apps/landing/src/components/navbar.tmp.tsx
@@ -14,7 +14,7 @@ import { useState } from "react";
 
 import { GitHubLogoIcon } from "@radix-ui/react-icons";
 import { Menu } from "lucide-react";
-import { ModeToggle } from "./mode-toggle";
+import { ModeToggle } from "./mode-toggle.tmp";
 import { buttonVariants } from "./ui/button";
 
 interface RouteProps {
